+++ Makefile
@@ -13,14 +13,15 @@ LIBFT = libft/libft.a
 
 NAME = minishell
 
-SRCS = src/minishell.c src/main_loop.c src/analize_line.c \
-		src/utils/error_message.c src/utils/utils.c \
+SRCS = src/utils/testcode.c \
+		src/minishell.c src/main_loop.c src/analize_line.c \
+		src/utils/error_message.c src/utils/setting.c \
 		src/signal/signal.c \
 		src/tokenizer/process_token.c src/tokenizer/memory.c \
 		src/tokenizer/tokenize.c src/tokenizer/tokenizer.c \
 		src/parser/parser.c src/parser/insert_tree.c src/parser/token_tool.c \
-		src/order/order_tree.c \
-		src/builtins/builtins.c src/builtins/echo.c src/builtins/pwd.c
+		src/exec_tree/order_tree.c \
+		src/builtins/builtins.c src/builtins/echo.c src/builtins/pwd.c 
 
 OBJS = $(SRCS:.c=.o)
