+++ src/builtins/func.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pwd.c                                              :+:      :+:    :+:   */
+/*   func.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jiwonhan <jiwonhan@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/19 00:13:28 by jiwonhan          #+#    #+#             */
-/*   Updated: 2022/12/21 17:24:06 by jiwonhan         ###   ########seoul.kr  */
+/*   Updated: 2022/12/21 20:12:09 by jiwonhan         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,6 @@
 void	cd(void)
 {
 	printf("cd\n");
-}
-
-void	pwd(void)
-{
-	printf("pwd\n");
 }
 
 void	ft_export(void)
