+++ src/main_loop.c
@@ -6,10 +6,11 @@
 /*   By: jiwonhan <jiwonhan@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 16:38:08 by jiwonhan          #+#    #+#             */
-/*   Updated: 2022/12/21 19:59:33 by jiwonhan         ###   ########seoul.kr  */
+/*   Updated: 2022/12/21 20:46:14 by jiwonhan         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "testcode.h"	//TODO
 #include "minishell.h"
 #include "tokenization.h"
 #include "parse_tree.h"
@@ -74,7 +75,6 @@ static int check_line(char **line)
 	{
 		*line = ft_strjoin(*line, "\n");
 		*line = ft_strjoin(*line, readline("> "));
-		//printf("join line = %s@\n", *line);
 		quote = check_quote(*line);
 	}
 	return (1);
@@ -99,25 +99,14 @@ void main_loop(t_minishell *minishell)
 	while (read_line(&line))
 	{
 		add_history(line);
-		check_line(&line); //TODO jiwon, false=> free
-		//printf("input: %s\n", line);	//TODO
+		check_line(&line);
 		tokenization = analize_line(minishell, line);	//TODO error처리를 위해 우선 minishell도 매개변수로 주었음
 		free(line);
 		if (!tokenization)
 			continue ;
-		/*
-		printf(">>tokenization result<<\n");	//TODO
-		t_token *token = tokenization;
-		while (token)
-		{
-			printf("arg: %10s type: %10d len: %10zu\n", token->value, token->type, ft_strlen(token->value));
-			if (token->type == 3)
-				printf("<======= $: %s ========>\n", process_dquote(token->value));
-			token = token->next;
-		}*/	//TODO tokenization result end
-		parse_tree = parser(tokenization);//(void)parse_tree;
-		//printf(">>parse tree result<<\n");	//TODO
-		//print_parse_tree(parse_tree, 0);
+		//print_tokenization_result(tokenization);	//TODO
+		parse_tree = parser(tokenization);
+		//(void)parse_tree;printf(">>parse tree result<<\n");print_parse_tree(parse_tree, 0);
 		exev_line(minishell, parse_tree);
 		free_tokens(parse_tree->token);
 		//TODO parse tree result end
@@ -125,29 +114,4 @@ void main_loop(t_minishell *minishell)
 		free (parse_tree);
 	}
 	return ;
-}
-
-void print_node(t_token *token)
-{
-	t_token *tmp;
-
-	tmp = token;
-	while (tmp)
-	{
-		printf("%s\t", tmp->value);
-		tmp = tmp->next;
-	}
-	printf("\n");
-}
-
-void print_parse_tree(t_parse_tree *parse_tree, int level)
-{
-	printf("level%d root(%d)\t", level, parse_tree->type);
-	print_node(parse_tree->token);
-	if (parse_tree->original)
-		printf("\toriginal : %s\n", parse_tree->original);
-	if (parse_tree->left)
-		print_parse_tree(parse_tree->left, level + 1);
-	if (parse_tree->right)
-		print_parse_tree(parse_tree->right, level + 1);
 }
