+++ src/minishell.c
@@ -6,7 +6,7 @@
 /*   By: jiwonhan <jiwonhan@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 16:38:12 by jiwonhan          #+#    #+#             */
-/*   Updated: 2022/12/19 03:11:32 by hyuncpar         ###   ########.fr       */
+/*   Updated: 2022/12/21 20:30:36 by jiwonhan         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	a()
 {
-	printf("?????\n");
+	printf("leaks test\n");
 	system("leaks minishell");
 }
 
@@ -23,9 +23,8 @@ int main(int ac, char *av[], char *envp[])
 	//atexit(a);
 	t_minishell	minishell;
 
-	//minishell = NULL;
 	check_arg(ac, av);
-	init(&minishell, envp);	//TODO hyuntall->envp
+	init(&minishell, envp);
 	main_loop(&minishell);
 	return (0);
 }
